us-620, t-638,639,641,642,654,655 Changes: Additional tables have been added to the db. These are in support of other user stories, such as us619. Updated pip requirements for docker environment. Added mock data for users. Tweaked the eslint config to be compatible with ecma2017.
Testing.
Docker must be running, and psycop2 must be installed.
From the proj root, perform a populated launch of the server.
```
pip install psycopg2-binary
./iron.sh -p
```

Open your browser to `localhost:3001/login` and enter `user2` and `password` and log in.
This should redirect you to the stand-in panel. Log out and go back to log in page. Enter `user1` and `password` and log in. It will not redirect to panel. Without closing the browser, go to `localhost:3001\data-entry` and you will see that you are logged in. There are currently 3 levels of user priviledge: admin, data-entry, and user. If the role is user or data-entry, login redirects, but won't for admin. Fixing this is outside the scope of this US. The mock data is a user for each role, user1 - user3. If you connect with DBViz, then you can inspect the data.