US173: Set default error & update title
Added logic to set a default error if the passed prop is undefined. Also
added some hooks to set the title to the given error code and unset the
title when the component is cleaned up.

In practice, this doesn't seem to work. The code executes as expected,
however, for whatever reason the title refuses to update on cleanup.